Added get_robot_io and get_electric_gripper_status functions
Added the two functions to read InOut_in array as well as Gripper_data_in. Currently InOut_in is only tested with E-Stop which works, and Gripper_data_in has non-working "object detected" value.
Also added a function to enable the user to input their COM port which is then saved into a .txt file so it is only needed to be done once.
Also made sure when E-Stop is pressed that the gripper is disabled.