Added smooth motion and command acknowledgment
Added a lot here, but the gist of it is an overhaul to the robot_api and headless_commander plus a smooth_motion.py (for the smooth motion logic). The code is completely backwards compatible, so you should not have any issues with current code. I will update the README as well with everything.